requirements.txt
azure-cosmos==4.14.3
azure-storage-blob==12.27.1
fastapi==0.127.0
fpdf==1.7.2
google-genai==1.56.0
langchain==1.2.0
langchain-chroma==1.1.0
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.2.5
langchain-google-genai==4.1.2
langchain-ollama==1.0.1
langchain-openai==1.1.6
langchain-postgres==0.0.16
langchain-text-splitters==1.1.0
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.1
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
uvicorn==0.40.0
